[rpc] allow ability to abort second call to RecvWork::wait() in ProcessGroupAgent::listenLoop
Pull Request resolved: https://github.com/pytorch/pytorch/pull/36084

Added support to abort the call to a `RecvWork` created by `recvAnysource` but there is an additional call to `pg_->recv()` to actually get the tensor sent over the wire (the previous call is the preamble for the tensor). This adds support to be able to abort this call as well in `::shutdown()`, which can be used to avoid hangs during ungraceful shutdown.

Added an internal test case in `ProcessGroupAgentTest` to ensure that an appropriate error message is raised when this happens.
ghstack-source-id: 101604969

Differential Revision: [D20632764](https://our.internmc.facebook.com/intern/diff/D20632764/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D20632764/)!